WIP: initial changes to make a client class for each workflow library
