+++ online_training/src/drivers/conf/ssim_config.yaml
