+++ online_training/src/drivers/ssim_driver.py
